extract `GitRemote`
